Instantiate - 옵션 예제별 패키지 정리
