+++ src/main/java/option/introstpector/constructor_properties/BothButAllArgsByManual.java
@@ -1,23 +1,19 @@
-package constructor_properties;
+package option.introstpector.constructor_properties;
 
 import jakarta.validation.constraints.NotBlank;
+import lombok.NoArgsConstructor;
 
 import java.beans.ConstructorProperties;
 
-public class NoArgsBeforeAllArgs {
+@NoArgsConstructor
+public class BothButAllArgsByManual {
     @NotBlank
     public String name;
     @NotBlank
     public String job;
 
-    public boolean onlyInitiatedByNoArgs;
-
-    public NoArgsBeforeAllArgs() {
-        this.onlyInitiatedByNoArgs = true;
-    }
-
     @ConstructorProperties({"name", "job"})
-    public NoArgsBeforeAllArgs(String name, String job) {
+    public BothButAllArgsByManual(String name, String job) {
         this.name = name;
         this.job = job;
     }
