+++ src/main/java/option/introstpector/constructor_properties/NoArgsBeforeAllArgs.java
@@ -1,10 +1,10 @@
-package constructor_properties;
+package option.introstpector.constructor_properties;
 
 import jakarta.validation.constraints.NotBlank;
 
 import java.beans.ConstructorProperties;
 
-public class AllArgsBeforeNoArgs {
+public class NoArgsBeforeAllArgs {
     @NotBlank
     public String name;
     @NotBlank
@@ -12,12 +12,13 @@ public class AllArgsBeforeNoArgs {
 
     public boolean onlyInitiatedByNoArgs;
 
+    public NoArgsBeforeAllArgs() {
+        this.onlyInitiatedByNoArgs = true;
+    }
+
     @ConstructorProperties({"name", "job"})
-    public AllArgsBeforeNoArgs(String name, String job) {
+    public NoArgsBeforeAllArgs(String name, String job) {
         this.name = name;
         this.job = job;
-    }
-    public AllArgsBeforeNoArgs() {
-        this.onlyInitiatedByNoArgs = true;
     }
 }
