+++ src/main/java/option/introstpector/constructor_properties/OnlyAllArg.java
@@ -1,13 +1,10 @@
-package constructor_properties;
-
+package option.introstpector.constructor_properties;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 
-@NoArgsConstructor
 @AllArgsConstructor
-public class NoArgsBeforeAllArgsByLomBock {
+public class OnlyAllArg {
     @NotBlank
     public String name;
     @NotBlank
